# lib/api-spec/openapi.yaml
openapi: 3.1.0
info:
  # Do not change the title, if the title changes, the import paths will be broken
  title: Api
  version: 0.1.0
  description: POWAPLAY API specification
servers:
  - url: /api
    description: Base API path
...
paths:
...
  /favorites/{projectId}:
    post:
      operationId: addFavorite
      tags: [favorites]
      summary: Favorite a project
      parameters:
        - $ref: '#/components/parameters/AuthorizationSessionHeader'
        - $ref: '#/components/parameters/SessionCookie'
        - name: projectId
          in: path
          required: true
          schema:
            type: integer
      responses:
        "200":
          description: Favorited
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/FavoriteResponse"
        "401":
          description: Not authenticated
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorEnvelope"
    delete:
      operationId: removeFavorite
      tags: [favorites]
      summary: Unfavorite a project
      parameters:
        - $ref: '#/components/parameters/AuthorizationSessionHeader'
        - $ref: '#/components/parameters/SessionCookie'
        - name: projectId
          in: path
          required: true
          schema:
            type: integer
      responses:
        "200":
          description: Unfavorited
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/FavoriteResponse"
        "401":
          description: Not authenticated
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorEnvelope"
  /me/favorites:
    get:
      operationId: getMyFavorites
      tags: [favorites]
      summary: List user's favorited projects
      parameters:
        - $ref: '#/components/parameters/AuthorizationSessionHeader'
        - $ref: '#/components/parameters/SessionCookie'
      responses:
        "200":
          description: Favorited projects
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ProjectListResponse"
        "401":
          description: Not authenticated
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorEnvelope"
...
components:
  parameters:
    AuthorizationSessionHeader:
      name: Authorization
      in: header
      required: false
      description: Opaque session token — Bearer <sid>.
      schema:
        type: string
    SessionCookie:
      name: sid
      in: cookie
      required: false
      description: Browser session cookie.
      schema:
        type: string
  schemas:
...
    ErrorEnvelope:
      type: object
      required: [error]
      properties:
        error:
          type: string
    ProjectDetail:
      type: object
      required: [id, title, slug, url, description, tags, style, isPublic, isHidden, thumbnailUrl, previewVideoUrl, favoriteCount, ownerUsername, ownerAvatarUrl, createdAt]
      properties:
        id:
          type: integer
        title:
          type: string
        slug:
          type: string
        url:
          type: string
        description:
          type: [string, 'null']
        tags:
          type: array
          items:
            type: string
        style:
          type: [string, 'null']
        isPublic:
          type: boolean
        isHidden:
          type: boolean
        thumbnailUrl:
          type: [string, 'null']
        previewVideoUrl:
          type: [string, 'null']
        favoriteCount:
          type: integer
        ownerUsername:
          type: string
        ownerAvatarUrl:
          type: [string, 'null']
        createdAt:
          type: string
          format: date-time
    ProjectListResponse:
      type: object
      required: [projects, total]
      properties:
        projects:
          type: array
          items:
            $ref: "#/components/schemas/ProjectDetail"
        total:
          type: integer
...
    FavoriteResponse:
      type: object
      required: [success]
      properties:
        success:
          type: boolean
...