# lib/api-zod/src/generated/types/projectDetail.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * POWAPLAY API specification
 * OpenAPI spec version: 0.1.0
 */

export interface ProjectDetail {
  id: number;
  title: string;
  slug: string;
  url: string;
  /** @nullable */
  description: string | null;
  tags: string[];
  /** @nullable */
  style: string | null;
  isPublic: boolean;
  isHidden: boolean;
  /** @nullable */
  thumbnailUrl: string | null;
  /** @nullable */
  previewVideoUrl: string | null;
  favoriteCount: number;
  ownerUsername: string;
  /** @nullable */
  ownerAvatarUrl: string | null;
  createdAt: Date;
}